Remove code for testing up/down measurement errors.
Determination: the "standard" implementation of moving the caret up or
down (i.e. that found in Chrome that I could not *quite* get my code to
replicate), DOES maintain a "memory" of where the caret was before. I
found instances where moving up and then down did NOT have the same
result as independently positioning the cursor on the higher line and
then pressing down. The final implementation will need to account for
this.