Refined next-word function.
Changed behavior of next-word function to better match expectations when
the starting character is a separator. It will now skip to the end of
the word in cases where the immediate next character is a word
character. This is useful in case of dashed-together words -- if the
cursor is in the dash, the behavior of most editors is to skip until the
end of the word. This also preserves the niceity of treating an ellipsis
as a separate word, instead of skipping all the way over it and to the
end of the next word (like Sublime Text does).

Also updated whitespace? function to work for all whitespace, not just a
single space character.