Reorg of navigation code, made Navigable protocol.
Switched next-word to take a sel instead of offset, and made it a part
of a Protocol called Navigable (which will also contain forward, back,
start, end, next-sentence, previous-sentence, etc). This will be
extended to handle Documents as well in time. Open polymorphism FTW.

Also added a skeleton for prev-word, but not yet implemented.