First-pass implementation of prev-word.
Initial implementation of prev-word done and appears to work, but I HAVE
NOT written tests yet, so there may be bugs. I added an reverse version
of the `until` helper func, `back-until`, along with helpers calling it.

Also fixed a bug in `next-word-offset` where it failed when the starting
index == the length of the string. Prevented the converse from happening
at the beginning of the string with `prev-word-offset`.