Wrote tests for prev-word.
Wrote test cases for prev-word, shifted some behavior of both next-word
and prev-word in response to a bug I found with handling inter-word
punctuation in prev-word. Now, both functions will no longer treat a
single punctuation mark inside a word as a separate word - time will
tell if this is the right approach, but I feel good about it.