Fix broken tests, and a bunch of other things.
This was a cascading series of minor bugs that followed on from cleanup
and fixing tests. Should really have been separate commits, but I would
prefer not to leave the codebase in a state where tests aren't passing.

Change list:

- Fix error with paragraph single delete at start (no test yet).
- Fix failure in at-offset and before-offset when passed an empty run.
- Fix failure with delete-after at end of paragraph (no test yet).
- Minor doc fixes and cleanup while poking around old code.
- Probably other things I'm forgetting.

Tests will be added on in subsequent commits, so be warned this is not
a good spot to revert back to.