Implement shift+right and shift+left functionality
- Added `Selectable` protocol (implemented on Document) inside of nav
namespace for handling shifting selections around.

- Modified some of the functions for expanding-contracting in the
selection namespace. There are now three functions: (shift-start) for
moving the start of the selection, (shift-end) for moving the end, and
(shift-caret) for shifting the *side that the caret is on*, defaulting
to the right side in the case of a single selection. Initially I thought
this might help with some edge cases, though it's not actually useful in
the case of shift+left. I might remove it in the future.

- Added a (caret-para) function that returns the UUID of the paragraph
a selection's caret is inside of