Fixed regression in down interceptor/nearest-line-offset-to-pixel
Introducing a fix to issue where clicking off the right hand side of a line would cause the cursor to be placed immediately after the end-of-line space (visually represented as being the very start of the next line) caused a regression when either clicking off the right bound of the _last_ line in a pargraph, or when navigating down into an empty paragraph (i.e. where :start-offset and :end-offset are both 0), either by clicking or with down arrow. Did this by introducing a new boolean parameter to nearest-line-offset-to-pixel indicating whether the line passed in is the last line in its paragraph.