Determine cause of bottleneck in click after select all
After performing a select all, each time click is fired, the changelist
is massive (all UUIDs in the document, basically). This is expected for
the first click interceptor called after select-all, as it has to
deselect everything to then reposition the caret. However, the
changelist is still massive for _every_ subsequent call to click, until
another entry is added to the :history object (e.g. by entering some
text and waiting for the timeout to lapse).

Most of this commit is just investigational code that will be deleted at
some point. There are some minor modifications to the
(inside-time-measurement!) macro as well, that will start the supplied
measurement if it has not already been started.