Remove DocumentFragment and ParagraphFragment.
Also changed the behavior around changing the type of a paragraph when
another paragraph is inserted into it; now paragraph A will only change
to type of paragraph B if paragraph A is empty, otherwise the type of
paragraph A remains the same. Previously, the type would be changed if
inserting into the beginning of paragraph B, which caused some
confusion, and the ability to have a series of runs without a type
associated with them was needed, hence why ParagraphFragment was
birthed.

However, I think that was more confusing than it was worth and added the
need to handle extra types to the system (essentially reimplementing
everything that already took a Paragraph to work with a
ParagraphFragment as well). Much easier to just use a Paragraph anytime
a group of runs is needed. If there's a need to have a group of runs
_explicitly with no paragraph type associated_, a :type of `nil` can
always be used for that purpose.